resolve runtime errors (platform)
