+++ .lintstagedrc.json
@@ -1,4 +1,4 @@
 {
-  "*.{js,jsx,ts,tsx}": ["eslint --fix", "prettier --write"],
-  "*.{json,md,css,scss}": ["prettier --write"]
+  "*.{js,mjs,cjs,jsx,ts,tsx}": ["eslint --fix", "prettier --write"],
+  "*.{html,json,md,css,scss}": ["prettier --write"]
 }
