+++ platform/.env.example
@@ -4,7 +4,7 @@ AUTH_ARGON2_MEMORY=19456
 AUTH_ARGON2_PARALLELISM=1
 AUTH_SESSION_COOKIE_NAME=svg_session
 AUTH_SESSION_TTL_HOURS=720
-DATABASE_URL="file:../data/sovereign.db"
+DATABASE_URL="file:../../data/sovereign.db"
 DEFAULT_USER_ROLE=platform:user
 FT_PROJECT_TYPE_BLOG=true
 FT_PROJECT_TYPE_PAPERTRAIL=true
