+++ platform/index.cjs
@@ -1,0 +1,4 @@
+require("dotenv").config();
+
+// TODO: use dist/ instead /src for production
+require("./src/bootstrap.js").bootstrap();
