+++ platform/src/config/flags.mjs
@@ -1,4 +1,4 @@
-import { toBool } from "$/utils/misc.mjs";
+import { toBool } from "../utils/misc.mjs";
 
 // TODO: Should allow registering new project types and flags.
 // TODO: Fetch feature toggles from DB-stored config (e.g. per-tenant or per-user overrides).
