+++ platform/src/platform/handlers/app.mjs
@@ -1,6 +1,6 @@
-import { prisma } from "$/services/database.mjs";
-import logger from "$/services/logger.mjs";
-import { refreshEnvCache } from "$/config/env.mjs";
+import { prisma } from "../../services/database.mjs";
+import logger from "../../services/logger.mjs";
+import { refreshEnvCache } from "../../config/env.mjs";
 
 const DEFAULT_SCOPE = "platform";
 const KEY_MAX_LENGTH = 200;
