+++ platform/src/platform/handlers/linkPreview.mjs
@@ -1,6 +1,6 @@
 import fetch from "node-fetch";
 
-import logger from "$/services/logger.mjs";
+import logger from "../../services/logger.mjs";
 
 export async function fetchLinkPreview(req, res) {
   try {
