+++ platform/src/platform/handlers/projects/core/update.mjs
@@ -1,6 +1,6 @@
-import { prisma } from "$/services/database.mjs";
-import logger from "$/services/logger.mjs";
-import { ensureProjectAccess, ProjectAccessError } from "$/utils/projectAccess.mjs";
+import { prisma } from "../../../../services/database.mjs";
+import logger from "../../../../services/logger.mjs";
+import { ensureProjectAccess, ProjectAccessError } from "../../../../utils/projectAccess.mjs";
 
 export default async function update(req, res) {
   try {
