+++ platform/src/platform/handlers/users/index.mjs
@@ -1,7 +1,7 @@
-import { prisma } from "$/services/database.mjs";
-import logger from "$/services/logger.mjs";
-import { syncProjectPrimaryOwner } from "$/utils/projectAccess.mjs";
-import { USER_ROLES } from "$/config/index.mjs";
+import { prisma } from "../../../services/database.mjs";
+import logger from "../../../services/logger.mjs";
+import { syncProjectPrimaryOwner } from "../../../utils/projectAccess.mjs";
+import { USER_ROLES } from "../../../config/index.mjs";
 
 const DATE_FORMAT = new Intl.DateTimeFormat("en-US", {
   dateStyle: "medium",
