+++ platform/src/platform/middlewares/auth.mjs
@@ -1,5 +1,9 @@
-import { getSessionWithUser, getOrCreateSingletonGuestUser, createSession } from "$/utils/auth.mjs";
-import env from "$/config/env.mjs";
+import {
+  getSessionWithUser,
+  getOrCreateSingletonGuestUser,
+  createSession,
+} from "../../utils/auth.mjs";
+import env from "../../config/env.mjs";
 
 const {
   AUTH_SESSION_COOKIE_NAME,
