--- platform/src/plugins/blog/plugin.json
@@ -1,109 +1,0 @@
-{
-  "namespace": "blog",
-  "name": "@sovereign/blog",
-  "version": "1.0.0-alpha.7",
-  "schemaVersion": 1,
-  "preRelease": true,
-  "isEnabled": true,
-  "description": "Sovereign Blog",
-  "main": "./index.mjs",
-  "author": "Sovereign Core Team",
-  "license": "AGPL-3.0",
-  "sovereign": {
-    "engine": "0.5.0",
-    "entryPoints": ["launcher"]
-  },
-  "routes": {
-    "web": true,
-    "api": true
-  },
-  "platformCapabilities": {
-    "database": true,
-    "gitManager": true,
-    "fs": true
-  },
-  "userCapabilities": [
-    {
-      "key": "user:plugin.blog.feature",
-      "description": "Enable Blog plugin.",
-      "roles": ["platform:user"]
-    },
-    {
-      "key": "user:plugin.blog.create",
-      "description": "Create blog project, and configure.",
-      "roles": ["platform:user"]
-    },
-    {
-      "key": "user:plugin.blog.read",
-      "description": "View own blog project.",
-      "roles": [
-        "project:admin",
-        "project:editor",
-        "project:contributor",
-        "project:viewer",
-        "project:guest"
-      ]
-    },
-    {
-      "key": "user:plugin.blog.update",
-      "description": "Update blog project.",
-      "roles": ["project:admin", "project:editor", "project:contributor"]
-    },
-    {
-      "key": "user:plugin.blog.update.shared",
-      "description": "Update blog project.",
-      "roles": ["project:admin", "project:editor", "project:contributor"]
-    },
-    {
-      "key": "user:plugin.blog.delete",
-      "description": "Delete blog project.",
-      "roles": ["project:admin"]
-    },
-    {
-      "key": "user:plugin.blog.post.create",
-      "description": "Create blog posts.",
-      "roles": ["project:admin", "project:editor", "project:contributor"]
-    },
-    {
-      "key": "user:plugin.blog.post.read",
-      "description": "View blog posts.",
-      "roles": [
-        "project:admin",
-        "project:editor",
-        "project:contributor",
-        "project:viewer",
-        "project:guest"
-      ]
-    },
-    {
-      "key": "user:plugin.blog.post.update",
-      "description": "Update blog posts.",
-      "roles": ["project:admin", "project:editor", "project:contributor"]
-    },
-    {
-      "key": "user:plugin.blog.post.update.shared",
-      "description": "Update shared blog posts.",
-      "roles": ["project:admin", "project:editor", "project:contributor"]
-    },
-    {
-      "key": "user:plugin.blog.post.delete",
-      "description": "Delete blog posts.",
-      "roles": ["project:admin", "project:editor", "project:contributor"]
-    },
-    {
-      "key": "user:plugin.blog.post.delete.shared",
-      "description": "Delete shared blog posts.",
-      "roles": ["project:admin", "project:editor"]
-    },
-    {
-      "key": "user:plugin.blog.post.publish",
-      "description": "Publish blog posts.",
-      "roles": ["platform:user", "project:admin", "project:editor"]
-    }
-  ],
-  "events": {},
-  "prisma": {},
-  "config": {
-    "FT_PLUGIN_BLOG": true
-  }
-}
